feat: Add comments app & update ERD
This commit will add the comments app to the project.
It will add the basic model. The ERD for Comment was updated
because I did add a comment field to id, which will probably lead to
confusion when called like comment.comment. Got renamed to content.
ERD is updated.
The base serializer is created.