feat: Add guns app to project
This commit will add the guns app to the project.
It will create the models, include a choice field for gun type and
provide the same default image as used before.

It creates an additional permissions class to make sure only the owner
has access to a new created object. This will harden the access to
prevent that another logged in user will ever have access to this
object.

It add a create, list view in order to create and list an object.