feat: Add CRUD to gun app
This commit will add CRUD functionality to the gun app.
It will create the detail view and bring the feature to update
and delete to it, if you are logged in and the owner.

It will add the serial_number field to thd model and it will adjust
the defautl sorting in the list view.