feat: Add update & delete feature to usermessages
This commit will add the update and delete feature to the usermessages
app.
Furthermore, it will create the message detail view.
It will create a new custom permission to check that only the sender
of a message can update and delete it. This was kind of necessary,
since the owner field does not exist in the usermessages model.