feat: Modify DATETIME, DATE Format
This commit will modify the DATETIME as well as the DATE format.
Two approaches are taken. For the "normal" DATETIME format, the
format like "13 Aug 2023" will be used, which is well readable.

However, for the searchable filter field like match_date, this does not
work, since the database values are searched, which are 13-08-2023.
In order not to confuse the user, 13 08 2023 is user for the format
instead.