feat: Adjust serializers date_time field
This commit will adjust the match_date field, when passed to the
serializers, to use the custom SerializerMethodField. In the
serializer, we have a new method, which translates the date_time
filed to the desried value of, for example, 13 Aug 2023
This will make our life in the frontend react part much easier,
when using filter / search options.