fix: Adjust match_date in matches serializers
The previously changed match_date code in the match serializer did
not work. It was no longer possible to create a match, because
of the invalid date settings.
The following adjustment should fix this problem. At least from
the API itself it is possible again to create a match. Frontend
needs to be checked / adjusted accordingly.